sleep sleep sleep

+++ 0x14-bit_manipulation/1-print_binary.c
@@ -7,10 +7,10 @@
  * Return: no shit
  */
 
-void print_binary(unsigned int n)
+void print_binary(unsigned long int n)
 {
 	if (n > 1)
 		print_binary(n >> 1);
 
-	printf("%d", n & 1);
+	printf("%lu", n & 1);
 }
